requirements.txt
addict==2.2.1
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
appdirs==1.4.4
argon2-cffi==20.1.0
asn1crypto==1.3.0
async-timeout==3.0.1
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.8.2
black==19.10b0
bleach==3.1.5
blessings==1.7
cachetools==4.2.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
conda==4.8.1
conda-build==3.18.9
conda-package-handling==1.6.0
contextvars==2.4
cryptography==2.8
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
easydict==1.9
entrypoints==0.3
filelock==3.0.12
future==0.18.2
glob2==0.7
google-api-core==1.30.0
google-auth==1.32.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
graphviz==0.8.4
grpcio==1.38.1
gym==0.18.0
h5py==2.8.0
h5py-cache==1.0
hiredis==2.0.0
idna==2.7
idna-ssl==1.1.0
imageio==2.9.0
immutables==0.15
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.10.3
joblib==0.16.0
jsonpatch==1.24
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
libarchive-c==2.8
lief==0.9.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
msgpack==1.0.2
multidict==5.1.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nose==1.3.7
notebook==6.1.0
numpy==1.18.1
nvidia-ml-py3==7.352.0
olefile==0.46
omegaconf==2.0.0
opencensus==0.7.13
opencensus-context==0.1.2
opencv-python==4.3.0.36
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
portalocker==2.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.17.3
psutil==5.6.7
ptyprocess==0.6.0
py-spy==0.3.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.19
pydantic==1.8.2
pygame==1.9.6
pyglet==1.5.0
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==18.1.1
qtconsole==4.7.5
QtPy==1.9.0
ray==1.4.0
redis==3.5.3
regex==2020.7.14
requests==2.19.1
rsa==4.7.2
ruamel-yaml==0.15.87
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.13.0
soupsieve==1.9.5
tensorboardX==2.1
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.7.24
toml==0.10.1
torch==1.6.0
torch-encoding==1.2.1
torchfile==0.1.0
torchnet==0.0.4
torchsummary==1.5.1
torchvision==0.5.0
tornado==6.0.3
tqdm==4.41.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.23
visdom==0.1.8.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.5.1
yarl==1.6.3
zipp==3.1.0
